pyproject.toml
[project]
name = "openwebui-rag-fastapi"
version = "0.1.0"
description = "FastAPI OpenAI-compatible endpoint for Open WebUI + Memvid RAG + Ollama"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "httpx>=0.27",
  "pyyaml>=6.0",
  "pypdf>=4.0",
  "memvid-sdk",
]
